Added JavaRuntime as core class;

+++ KD.Dova/Natives/JavaVMAttachArgs.cs
@@ -14,7 +14,7 @@ namespace KD.Dova.Proxy.Natives
 {
     internal unsafe struct JavaVMAttachArgs
     {
-        public IntPtr version;
+        public int version;
         public IntPtr name;
         public IntPtr group;
     }
